bug fixes, added some comments, modularized, and updated expand all to be consistent with column filters
