+++ src/onLayout/customizeCheckboxes.js
@@ -4,7 +4,7 @@ export default function customizeCheckboxes() {
     //Redefine change event listener of Expand All checkbox.
     this.controls.wrap
         .selectAll('.control-group')
-        .filter(d => d.type === 'checkbox')
+        .filter(d => d.option === 'expand_all')
         .select('.changer')
         .on('change', function(d) {
             var changer_this = this;
@@ -38,8 +38,18 @@ export default function customizeCheckboxes() {
                         context.config[d.option] = changer_this.checked;
 
                         if (changer_this.checked) {
-                            context.draw(context.data.summarized);
+                            context.data.raw = context.data.summarized;
+                            // need to filter rows when expanding in case some input boxes are in use
+                            if (context.columnControls.filtered) {
+                                context.data.raw = context.data.raw.filter(
+                                    f => !f.filtered || f.visible_child
+                                );
+                            }
+                            context.draw(context.data.raw);
                             context.expandable_rows.classed('chm-table-row--collapsed', false);
+                            // I'm making the default when the chart is drawn to collapse all rows and (have the box unchecked)
+                            // however I do want it to be checked when it's supposed to so flipping it back here
+                            changer_this.checked = context.config[d.option];
                         } else {
                             context.draw(context.data.top);
                             context.expandable_rows.classed('chm-table-row--collapsed', true);
