+++ src/onLayout/customizeFilters/redraw.js
@@ -3,7 +3,8 @@ import resetFilters from '../addColumnControls/addResetButton/resetFilters';
 
 export default function redraw() {
     summarizeData.call(this);
-    this.data.raw = this.data.summarized.filter(d => d.parents.length == 0);
+    this.data.top = this.data.summarized.filter(d => d.parents.length == 0);
+    this.data.raw = this.data.top;
     resetFilters.call(this);
     this.draw(this.data.raw);
 }
